ICU-22675 Migrate from deprecated boxed primitive constructors to their replacements

+++ icu4j/main/core/src/main/java/com/ibm/icu/impl/PropsVectors.java
@@ -378,7 +378,7 @@ public class PropsVectors {
         // sort the properties vectors to find unique vector values
         Integer[] indexArray = new Integer[rows];
         for (int i = 0; i < rows; ++i) {
-            indexArray[i] = Integer.valueOf(columns * i);
+            indexArray[i] = columns * i;
         }
 
         Arrays.sort(indexArray, new Comparator<Integer>() {
